refactor(dashboard): consolidate utility functions and enhance settings management
- Extract storage key prefixing logic to shared utils module for reuse across components
- Move computeRenderableSettings import to centralized utils file
- Add TypeScript 6.0.3 to dev dependencies for improved type checking
- Update clipboard component to use shared getPrefixedKey utility
- Enhance files component to listen for serverSettings messages and conditionally render upload/download buttons based on settings
- Refactor settings component to use centralized utility imports
- Replace direct fullscreen API call with postMessage for better iframe compatibility
- Consolidate per-display settings and storage key logic into reusable utilities
- Improve type safety and reduce code duplication across dashboard components
