+++ addons/selkies-dashboard-wish/package.json
@@ -51,6 +51,7 @@
     "postcss": "^8.5.14",
     "tailwindcss": "^4.2.4",
     "tw-animate-css": "^1.4.0",
+    "typescript": "^6.0.3",
     "vite": "^8.0.11",
     "vite-plugin-minify": "^2.1.0"
   }
