+++ addons/selkies-dashboard-wish/src/components/ui/badge.tsx
@@ -19,6 +19,8 @@ const badgeVariants = cva(
         ghost:
           "hover:bg-muted hover:text-muted-foreground dark:hover:bg-muted/50",
         link: "text-primary underline-offset-4 hover:underline",
+        green:
+          "bg-green-500/10 text-green-600 dark:bg-green-500/20 dark:text-green-400",
       },
     },
     defaultVariants: {
